requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Babel
python-dotenv
psycopg2-binary
werkzeug